requirements/dev.txt
-r base.txt
django-debug-toolbar==4.0
django-extensions==3.2
ipython==8.0
